[step1] 비즈니스 로직 리팩터링 완료
- 패키지 구조 변경
- Section 이 메인이 아닌 Line 에서 비즈니스 로직을 실행하도록 코드 변경
