+++ src/main/java/nextstep/subway/line/domain/LineRepository.java
@@ -1,5 +1,6 @@
-package nextstep.subway.line;
+package nextstep.subway.line.domain;
 
+import nextstep.subway.line.domain.Line;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
