+++ src/main/java/nextstep/subway/line/dto/LineInfoDto.java
@@ -1,4 +1,4 @@
-package nextstep.subway.line;
+package nextstep.subway.line.dto;
 
 import lombok.Getter;
