+++ src/main/java/nextstep/subway/line/dto/LineStationsResponse.java
@@ -1,4 +1,4 @@
-package nextstep.subway.line;
+package nextstep.subway.line.dto;
 
 import lombok.AccessLevel;
 import lombok.EqualsAndHashCode;
