+++ src/main/java/nextstep/subway/line/dto/SectionRequest.java
@@ -1,4 +1,4 @@
-package nextstep.subway.section;
+package nextstep.subway.line.dto;
 
 import lombok.AccessLevel;
 import lombok.Getter;
