--- src/main/java/nextstep/subway/section/SectionRepository.java
@@ -1,6 +1,0 @@
-package nextstep.subway.section;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface SectionRepository extends JpaRepository<Section, Long> {
-}
