+++ src/test/java/nextstep/subway/line/domain/LineTest.java
@@ -1,6 +1,6 @@
-package nextstep.subway.line;
+package nextstep.subway.line.domain;
 
-import nextstep.subway.section.SectionFixtures;
+import nextstep.subway.line.SectionFixtures;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
