+++ src/test/java/nextstep/subway/unit/LineServiceTest.java
@@ -1,11 +1,11 @@
 package nextstep.subway.unit;
 
-import nextstep.subway.line.Line;
-import nextstep.subway.line.LineRepository;
-import nextstep.subway.line.LineResponse;
-import nextstep.subway.line.LineService;
-import nextstep.subway.section.Section;
-import nextstep.subway.section.SectionRequest;
+import nextstep.subway.line.domain.Line;
+import nextstep.subway.line.domain.LineRepository;
+import nextstep.subway.line.dto.LineResponse;
+import nextstep.subway.line.service.LineService;
+import nextstep.subway.line.domain.Section;
+import nextstep.subway.line.dto.SectionRequest;
 import nextstep.subway.station.Station;
 import nextstep.subway.station.StationFixtures;
 import nextstep.subway.station.StationRepository;
